fix: change prices sent to ecommerce back to string

See merge request dev_tls/tlspay/payment-api!624

+++ app/Services/RefundItemsService.php
@@ -85,7 +85,7 @@ class RefundItemsService
                 'ri_xref_ti_id' => $rItem['ri_xref_ti_id'],
                 'ri_id' => $rItem['ri_id'],
                 'ri_quantity' => $rItem['ri_quantity'],
-                'ri_amount' => (float) number_format((float) $rItem['ri_amount'], 2, '.', ''),
+                'ri_amount' => number_format((float) $rItem['ri_amount'], 2, '.', ''),
                 'ri_reason_type' => $rItem['ri_reason_type'],
                 'ri_status' => $rItem['ri_status'],
                 'ri_invoice_path' => $rItem['ri_invoice_path'],
@@ -147,10 +147,10 @@ class RefundItemsService
                 'price_rule' => $service['ti_price_rule'],
                 'sku' => $service['ti_fee_type'],
                 'product_name' => $service['ti_fee_name'],
-                'price' => (float) number_format((float) $service['ti_amount'], 2, '.', ''),
+                'price' => number_format((float) $service['ti_amount'], 2, '.', ''),
                 'vat' => $service['ti_vat'],
                 'quantity' => $service['ti_quantity'],
-                'price_without_tax' => (float) number_format(
+                'price_without_tax' => number_format(
                     (float) $service['ti_amount'] - ($service['ti_vat'] / 100 * $service['ti_amount']),
                     2,
                     '.',
